requirements.txt
# Testing dependencies
behave>=1.2.6
PyHamcrest>=2.0.4